Refactor JSON parser
