+++ JSON.h
@@ -12,7 +12,6 @@ class JSON
 
 private:
     static void Validator(const std::string&);
-    static std::string Trim(const std::string&);
     static bool isNumber(const std::string&);
 
 public:
